refactor Pin 생성 역할 위임(Frame -> Ready, Pending)
