+++ src/main/java/bowling/domain/status/running/Ready.java
@@ -1,5 +1,6 @@
 package bowling.domain.status.running;
 
+import bowling.domain.frame.Pins;
 import bowling.domain.status.Running;
 import bowling.domain.status.Status;
 import bowling.domain.status.finished.Strike;
@@ -8,11 +9,13 @@ public class Ready extends Running {
 
     public static final int STRIKE_SCORE = 10;
 
-    public Ready() {}
+    public Ready() {
+    }
 
     @Override
     public Status bowl(int downPin) {
-        if (downPin == STRIKE_SCORE) {
+        Pins pins = Pins.down(downPin);
+        if (pins.getDownPin() == STRIKE_SCORE) {
             return new Strike();
         }
         return new Pending(downPin);
